+++ src/cgeo/geocaching/cgData.java
@@ -1930,7 +1930,7 @@ public class cgData {
 		if (cursor.isNull(index)) {
 			cache.direction = null;
 		} else {
-			cache.direction = (Double) cursor.getDouble(index);
+			cache.direction = cursor.getFloat(index);
 		}
 		index = cursor.getColumnIndex("distance");
 		if (cursor.isNull(index)) {
